Move LoggedInUserTypes enum to types file
