+++ src/utils/localStorageUtils.ts
@@ -1,5 +1,5 @@
 import { AuthorizationTypes } from 'shared/constants';
-import { LoggedInUserTypes } from 'stores/main/utils';
+import { LoggedInUserTypes } from 'stores/main/types';
 
 export function removeUserFromStorage() {
   localStorage.removeItem('user');
